re-arrange package
