+++ mining/tensority/cgo_algorithm/lib/BytomPoW.h
