+++ mining/tensority/cgo_algorithm/lib/byte_order-allInOne.h
