+++ mining/tensority/go_algorithm/algorithm.go
@@ -1,0 +1,11 @@
+package go_algorithm
+
+import (
+	"github.com/bytom/protocol/bc"
+)
+
+func LegacyAlgorithm(bh, seed *bc.Hash) *bc.Hash {
+	cache := calcSeedCache(seed.Bytes())
+	data := mulMatrix(bh.Bytes(), cache)
+	return hashMatrix(data)
+}
