+++ mining/tensority/go_algorithm/seed.go
@@ -1,4 +1,4 @@
-package tensority
+package go_algorithm
 
 import (
 	"encoding/binary"
